Renamed index.js to sparkpost.js
