+++ lib/SendGridCompatibility/index.js
@@ -1,7 +1,7 @@
 'use strict';
 var _ = require('lodash');
 var url = require('url');
-var SparkPost = require('../index');
+var SparkPost = require('../sparkpost');
 
 /**
  * SendGrid compatibility constructor
