+++ test/spec/sparkpost.spec.js
@@ -3,7 +3,7 @@ var chai = require('chai')
   , sinon = require('sinon')
   , sinonChai = require('sinon-chai')
   , nock = require('nock')
-  , SparkPost = require('../../lib/index');
+  , SparkPost = require('../../lib/sparkpost');
 
 chai.use(sinonChai);
